% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cell_salts.R
\docType{data}
\name{cell_salts}
\alias{cell_salts}
\title{27 Human Cell Salts Table}
\format{
A data.table data frame with 27 rows and 2 variables:
\describe{
\item{Cell Salt}{The cell salt English names as used in homeopathy}
\item{Molecular Formula}{The molecular formula of the cell salt}
}
}
\source{
Irucka Ajani Embry, "Table of 27 Cell Salts/Tissue Remedies for Human Beings", \emph{Tuning to the Soil-ed Seed Vibrations For Our Collective Health}, Nashville, Tennessee: Questioning the Universe Publishing, 2024, page 13. See \url{https://www.qup.quest/soiled-seeds.html}.
}
\usage{
cell_salts
}
\description{
A table containing the 27 inorganic substances, known as
cell salts, found in Human Beings. This data set contains
2 of the 4 columns created for the "Table of 27 Cell
Salts/Tissue Remedies for Human Beings".
}
\keyword{datasets}
