% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cell_salts_atoms_final.R
\docType{data}
\name{cell_salts_atoms_final}
\alias{cell_salts_atoms_final}
\title{Total Number of Atoms of Elements From the 27 Human Cell Salts Table}
\format{
A data.table data frame with 24 rows and 3 variables:
\describe{
\item{Element Name}{The chemical element name}
\item{Element Symbol}{The chemical element symbol}
\item{Number of Atoms}{The total number of atoms contained within all 27 cell salts for this particular element}
}
}
\source{
\enumerate{
Kim S, Chen J, Cheng T, Gindulyte A, He J, He S, Li Q, Shoemaker BA, Thiessen PA, Yu B, Zaslavsky L, Zhang J, Bolton EE. PubChem 2023 update. Nucleic Acids Res. 2023 Jan 6;51(D1):D1373-D1380. doi:10.1093/nar/gkac956. PMID: 36305812, \url{https://pubchem.ncbi.nlm.nih.gov/}.
Stack Overflow. Question: "R - Total sum of all named items in list elements", \url{https://stackoverflow.com/questions/78450114/r-total-sum-of-all-named-items-in-list-elements}.
}
}
\usage{
cell_salts_atoms_final
}
\description{
A table containing the total number of atoms of the 24 chemical elements in
the 27 inorganic substances, known as cell salts, found in Human Beings.
}
\keyword{datasets}
