% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-civilian_nuclear_accidents_wiki.R
\docType{data}
\name{civilian_nuclear_accidents_wiki}
\alias{civilian_nuclear_accidents_wiki}
\title{Civilian nuclear accidents (Wikipedia)}
\format{
A data.table data frame with 28 rows and 5 variables:
\describe{
\item{Date}{Date of the accident}
\item{Location}{Location of the accident}
\item{INES Level}{INES level}
\item{Type}{Type of accident}
\item{Description}{Description of accident or incident}
}
}
\source{
Wikimedia Foundation, Inc. Wikipedia, 3 September 2023, "List of civilian nuclear accidents", \url{https://en.wikipedia.org/wiki/List_of_civilian_nuclear_accidents}.
}
\usage{
civilian_nuclear_accidents_wiki
}
\description{
A table containing the civilian nuclear accidents.
}
\keyword{datasets}
