% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cn_arid_semiarid_notes.R
\docType{data}
\name{cn_arid_semiarid_notes}
\alias{cn_arid_semiarid_notes}
\title{Notes for Table 2-2d: Runoff curve numbers for arid and semiarid rangelands}
\format{
A data.table data frame with 5 rows and 2 variables:
\describe{
\item{Note Number (*)}{The note numbers}
\item{Notes}{The notes}
}
}
\source{
United States Department of Agriculture Natural Resources Conservation Service Conservation Engineering Division, "Urban Hydrology for Small Watersheds Technical Release 55 (TR-55)", June 1986, \url{https://web.archive.org/web/20230810204711/https://directives.sc.egov.usda.gov/OpenNonWebContent.aspx?content=22162.wba} [Recovered with the Internet Archive: Wayback Machine]
}
\usage{
cn_arid_semiarid_notes
}
\description{
A table containing the notes for the curve numbers for arid and semiarid
areas.
}
\keyword{datasets}
