% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cn_other_agricultural.R
\docType{data}
\name{cn_other_agricultural}
\alias{cn_other_agricultural}
\title{Table 2-2c: Runoff curve numbers for other agricultural lands}
\format{
A data.table data frame with 19 rows and 7 variables:
\describe{
\item{Cover type}{Type of agricultural ground cover}
\item{Hydrologic condition *3}{Hydrologic condition}
\item{Curve numbers for hydrologic soil group A}{Curve numbers for soil group A}
\item{Notes}{Any notes for soil group A}
\item{Curve numbers for hydrologic soil group B}{Curve numbers for soil group B}
\item{Curve numbers for hydrologic soil group C}{Curve numbers for soil group C}
\item{Curve numbers for hydrologic soil group D}{Curve numbers for soil group D}
}
}
\source{
United States Department of Agriculture Natural Resources Conservation Service Conservation Engineering Division, "Urban Hydrology for Small Watersheds Technical Release 55 (TR-55)", June 1986, \url{https://web.archive.org/web/20230810204711/https://directives.sc.egov.usda.gov/OpenNonWebContent.aspx?content=22162.wba} [Recovered with the Internet Archive: Wayback Machine]
}
\usage{
cn_other_agricultural
}
\description{
A table containing curve numbers for other agricultural areas.
}
\keyword{datasets}
