% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cn_urban_notes.R
\docType{data}
\name{cn_urban_notes}
\alias{cn_urban_notes}
\title{Notes for Table 2-2a: Runoff curve numbers for urban areas}
\format{
A data.table data frame with 5 rows and 2 variables:
\describe{
\item{Note Number (*)}{The note numbers}
\item{Notes}{The notes}
}
}
\source{
United States Department of Agriculture Natural Resources Conservation Service Conservation Engineering Division, "Urban Hydrology for Small Watersheds Technical Release 55 (TR-55)", June 1986, \url{https://directives.sc.egov.usda.gov/OpenNonWebContent.aspx?content=22162.wba}
}
\usage{
cn_urban_notes
}
\description{
A table containing the notes for the curve numbers for urban areas.
}
\keyword{datasets}
