% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-common_gases_simp.R
\docType{data}
\name{common_gases_simp}
\alias{common_gases_simp}
\title{Common Gaseous Elements and Compounds Simplified}
\format{
A data.table data frame with 35 rows and 2 variables:
\describe{
\item{Name}{Name of the element or compound}
\item{Formula}{Molecular formula of the element or compound}
}
}
\source{
\enumerate{
   \item Chemistry LibreTexts, "Chapter 10.1: Gaseous Elements and Compounds", \url{https://chem.libretexts.org/LibreTexts/Howard_University/General_Chemistry\%3A_An_Atoms_First_Approach/Unit_4\%3A__Thermochemistry/Chapter_10\%3A_Gases/Chapter_10.1\%3A_Gaseous_Elements_and_Compounds}.
   \item Gaseous composition of dry air, \url{https://web.archive.org/web/20161222023409/https://eesc.columbia.edu/courses/ees/slides/climate/table_1.html}. Retrieved thanks to the Internet Archive: Wayback Machine. Original data source: Mackenzie, F.T. and J.A. Mackenzie (1995) \emph{Our changing planet}. Prentice-Hall, Upper Saddle River, NJ, p 288-307. (After Warneck, 1988; Anderson, 1989; Wayne, 1991.)
}
}
\usage{
common_gases_simp
}
\description{
A table containing common substances that are gases at 25 degrees C and 1.0
atm. This is the simplified table with the elements and compounds combined
in a column.
}
\keyword{datasets}
