% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dry_air.R
\docType{data}
\name{dry_air}
\alias{dry_air}
\title{Gaseous Composition of Dry Air}
\format{
A data.table data frame with 16 rows and 3 variables:
\describe{
\item{Name}{Name of the gas}
\item{Formula}{Chemical formula of the gas}
\item{Mole percent}{Fraction mole percent of the gas in dry air}
}
}
\source{
Gaseous composition of dry air, \url{https://web.archive.org/web/20161222023409/https://eesc.columbia.edu/courses/ees/slides/climate/table_1.html}. Retrieved thanks to the Internet Archive: Wayback Machine. Original data source: Mackenzie, F.T. and J.A. Mackenzie (1995) \emph{Our changing planet}. Prentice-Hall, Upper Saddle River, NJ, p 288-307. (After Warneck, 1988; Anderson, 1989; Wayne, 1991.)
}
\usage{
dry_air
}
\description{
A table containing the gaseous composition of dry air by mole percent.
}
\keyword{datasets}
