% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-exceptional_tn_waters_abbrv.R
\docType{data}
\name{exceptional_tn_waters_abbrv}
\alias{exceptional_tn_waters_abbrv}
\title{Abbreviations Used in the Known Exceptional Tennessee Waters and Outstanding National Resource Waters}
\format{
A data.table data frame with 17 rows and 2 variables:
\describe{
\item{Abbreviation}{Abbreviation used in the Known Exceptional Tennessee Waters and Outstanding National Resource Waters table}
\item{Full Name}{Full name for the abbreviation}
}
}
\source{
Tennessee Department of Environment and Conservation (TDEC) Division of Water Resources (DWR), Accessed 8 May 2024, The Known Exceptional Tennessee Waters and Outstanding National Resource Waters. \url{https://archive.fo/cCgd8}. Used the archive.today webpage capture for this URL for CRAN acceptance.
}
\usage{
exceptional_tn_waters_abbrv
}
\description{
A table containing the abbreviations used in the exceptional or outstanding
national resource waters in Tennessee table.
}
\keyword{datasets}
