% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Fp.R
\docType{data}
\name{Fp}
\alias{Fp}
\title{Table 4-2: Adjustment factor (Fp) for pond and swamp areas that are spread throughout the watershed}
\format{
A data.table data frame with 5 rows and 2 variables:
\describe{
\item{Percentage of pond and swamp areas}{Percent of pond and swamp areas}
\item{Fp}{Adjustment factor}
}
}
\source{
United States Department of Agriculture Natural Resources Conservation Service Conservation Engineering Division, "Urban Hydrology for Small Watersheds Technical Release 55 (TR-55)", June 1986, \url{https://web.archive.org/web/20230810204711/https://directives.sc.egov.usda.gov/OpenNonWebContent.aspx?content=22162.wba} [Recovered with the Internet Archive: Wayback Machine]
}
\usage{
Fp
}
\description{
A table containing the adjustment factor for percent of pond and swamp areas.
}
\keyword{datasets}
