% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/data-Fukushima_2011_FieldMeasurements_5_Metadata.R
\docType{data}
\name{Fukushima_2011_FieldMeasurements_5_Metadata}
\alias{Fukushima_2011_FieldMeasurements_5_Metadata}
\title{US DOE/NNSA and DoD Response to 2011 Fukushima Incident: Field Team Radiological Measurements Metadata}
\format{
A list with 21 objects
}
\source{
United States (US) Department of Energy (DOE)/National Nuclear Security Administration (NNSA), "US DOE/NNSA and DoD Response to 2011 Fukushima Incident: Field Team Radiological Measurements", \url{https://web.archive.org/web/20160617002257/https://catalog.data.gov/dataset/us-doe-nnsa-and-dod-response-to-2011-fukushima-incident-field-team-radiological-measuremen-33914}. Retrieved thanks to the Internet Archive: Wayback Machine
}
\usage{
Fukushima_2011_FieldMeasurements_5_Metadata
}
\description{
A list containing the metadata for US DOE/NNSA and DoD Response to 2011
Fukushima Incident: Field Team Radiological Measurements.
}
\keyword{datasets}
