% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/data-Fukushima_2011_FieldSampleAirResults_2_Metadata.R
\docType{data}
\name{Fukushima_2011_FieldSampleAirResults_2_Metadata}
\alias{Fukushima_2011_FieldSampleAirResults_2_Metadata}
\title{US DOE/NNSA and DoD Response to 2011 Fukushima Incident: Radiological Air Samples Metadata}
\format{
A list with 20 objects
}
\source{
United States (US) Department of Energy (DOE)/National Nuclear Security Administration (NNSA), "US DOE/NNSA and DoD Response to 2011 Fukushima Incident: Radiological Air Samples", \url{https://web.archive.org/web/20201108073914/https://catalog.data.gov/dataset/us-doe-nnsa-and-dod-response-to-2011-fukushima-incident-radiological-air-samples-09f5e}. Retrieved thanks to the Internet Archive: Wayback Machine
}
\usage{
Fukushima_2011_FieldSampleAirResults_2_Metadata
}
\description{
A list containing the metadata for US DOE/NNSA and DoD Response to 2011
Fukushima Incident: Radiological Air Samples.
}
\keyword{datasets}
