% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/data-Fukushima_2011_FieldSampleInstrumentResults_Metadata.R
\docType{data}
\name{Fukushima_2011_FieldSampleInstrumentResults_Metadata}
\alias{Fukushima_2011_FieldSampleInstrumentResults_Metadata}
\title{US DOE/NNSA Response to 2011 Fukushima Incident: Instrument Samples (InSitu Measurements) Metadata}
\format{
A list with 20 objects
}
\source{
United States (US) Department of Energy (DOE)/National Nuclear Security Administration (NNSA), "US DOE/NNSA Response to 2011 Fukushima Incident: Instrument Samples (InSitu Measurements)", \url{https://web.archive.org/web/20160630225626/https://catalog.data.gov/dataset/us-doe-nnsa-response-to-2011-fukushima-incident-instrument-samples-insitu-measurements-09ee4}. Retrieved thanks to the Internet Archive: Wayback Machine
}
\usage{
Fukushima_2011_FieldSampleInstrumentResults_Metadata
}
\description{
A list containing the metadata for US DOE/NNSA Response to 2011 Fukushima
Incident: Instrument Samples (InSitu Measurements).
}
\keyword{datasets}
