% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gas_constant.R
\docType{data}
\name{gas_constant}
\alias{gas_constant}
\title{Gas Constant in Various Units}
\format{
A data.table data frame with 26 rows and 2 variables:
\describe{
\item{R Values}{Gas constant (R) values}
\item{Units}{Unit for the R value}
}
}
\source{
Wikimedia Foundation, Inc. Wikipedia, 4 January 2022, "Gas constant", \url{https://en.wikipedia.org/wiki/Gas_constant}.
}
\usage{
gas_constant
}
\description{
A table containing the gas constant with various units.
}
\keyword{datasets}
