% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gases_plain.R
\docType{data}
\name{gases_plain}
\alias{gases_plain}
\title{Gases at Standard Temperature and Pressure (STP) {Wikipedia} -- Plain}
\format{
A data.table data frame with 540 rows and 6 variables:
\describe{
\item{Name}{Name of the gas}
\item{Formula}{Chemical formula of the gas}
\item{Boiling pt (°C)}{Boiling point of the gas in degrees Celsius}
\item{Melting pt (°C)}{Melting point of the gas in degrees Celsius}
\item{Molecular weight}{Molecular weight of the gas}
\item{CAS}{CAS number}
}
}
\source{
Wikimedia Foundation, Inc. Wikipedia, 26 August 2023, "List of gases", \url{https://en.wikipedia.org/wiki/List_of_gases}.
}
\usage{
gases_plain
}
\description{
A table containing gases at standard conditions without the brackets next to
the gas name in the Name column.
}
\keyword{datasets}
