% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-greenhouse_gases_cloudy_notes_wikipedia.R
\docType{data}
\name{greenhouse_gases_cloudy_notes_wikipedia}
\alias{greenhouse_gases_cloudy_notes_wikipedia}
\title{Greenhouse Gases -- Percent Contribution to Total Greenhouse Effect Notes (Wikipedia)}
\format{
A data.table data frame with 3 rows and 2 variables:
\describe{
\item{Note Number (*)}{The note numbers}
\item{Notes}{The notes}
}
}
\source{
Wikimedia Foundation, Inc. Wikipedia, 25 August 2023, "Greenhouse gas", \url{https://en.wikipedia.org/wiki/Greenhouse_gas}.
}
\usage{
greenhouse_gases_cloudy_notes_wikipedia
}
\description{
A table containing the notes associated with the greenhouse gas contribution table.
}
\keyword{datasets}
