% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-greenhouse_gases_cloudy_wikipedia.R
\docType{data}
\name{greenhouse_gases_cloudy_wikipedia}
\alias{greenhouse_gases_cloudy_wikipedia}
\title{Greenhouse Gases -- Percent Contribution to Total Greenhouse Effect (Wikipedia)}
\format{
A data.table data frame with 6 rows and 5 variables:
\describe{
\item{Contributor}{Greenhouse gas contributor name}
\item{K&T (1997) – Clear Sky}{Greenhouse gas individual contribution with a clear sky according to K&T}
\item{K&T (1997) – With Clouds}{Greenhouse gas individual contribution with a cloudy sky according to K&T}
\item{Schmidt (2010) – Clear Sky}{Greenhouse gas individual contribution with a clear sky according to Schmidt}
\item{Schmidt (2010) – With Clouds}{Greenhouse gas individual contribution with a cloudy sky according to Schmidt}
}
}
\source{
Wikimedia Foundation, Inc. Wikipedia, 25 August 2023, "Greenhouse gas", \url{https://en.wikipedia.org/wiki/Greenhouse_gas}.
}
\usage{
greenhouse_gases_cloudy_wikipedia
}
\description{
A table containing the greenhouse gas contribution table.
}
\keyword{datasets}
