% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-hsg_definitions.R
\docType{data}
\name{hsg_definitions}
\alias{hsg_definitions}
\title{Table from Appendix A: Definitions of Hydrologic Soil Groups (HSGs)}
\format{
A data.table data frame with 4 rows and 2 variables:
\describe{
item{Definitions}{Definition of the hydrologic soil groups}
item{Drainage}{Drainage information for the hydrologic soil groups, if any}
}
}
\source{
United States Department of Agriculture Natural Resources Conservation Service Conservation Engineering Division, "Urban Hydrology for Small Watersheds Technical Release 55 (TR-55)", June 1986, \url{https://directives.sc.egov.usda.gov/OpenNonWebContent.aspx?content=22162.wba}
}
\usage{
hsg_definitions
}
\description{
A table containing the definitions of the hydrologic soil groups.
}
\keyword{datasets}
