% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-i1_50.R
\docType{data}
\name{i1_50}
\alias{i1_50}
\title{1 1/2 Percent Effective Interest Table (Engineering Economy)}
\format{
A data.table data frame with 34 rows and 10 variables:
\describe{
\item{n}{The "number of interest periods"}
\item{F/P}{Future value given present value}
\item{P/F}{Present value given future value}
\item{F/A}{Future value given annual value}
\item{P/A}{Present value given annual value}
\item{A/F}{Annual value given future value}
\item{A/P}{Annual value given present value}
\item{P/G}{Present value given gradient}
\item{A/G}{Annual value given gradient}
\item{n}{The "number of interest periods"}
}
}
\source{
\enumerate{
   \item William G. Sullivan, Elin M. Wicks, and C. Patrick Koelling, \emph{Engineering Economy}, Fourteenth Edition, Upper Saddle River, New Jersey: Pearson/Prentice Hall, 2009, page 625-647.
   \item Masoud Olia, Ph.D., P.E. and Contributing Authors, \emph{Barron's FE (Fundamentals of Engineering Exam)}, 3rd Edition, Hauppauge, New York: Barron's Educational Series, Inc., 2015, page 178-187.
}
}
\usage{
i1_50
}
\description{
A table containing the number of periods and the corresponding engineering
economic value given the "effective interest rate" and period.
}
\keyword{datasets}
