% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Ia.R
\docType{data}
\name{Ia}
\alias{Ia}
\title{Table 4-1: Ia values for runoff Curve number}
\format{
A data.table data frame with 59 rows and 2 variables:
\describe{
\item{Curve number}{Runoff curve number}
\item{Ia (in)}{Initial abstraction in inches}
}
}
\source{
United States Department of Agriculture Natural Resources Conservation Service Conservation Engineering Division, "Urban Hydrology for Small Watersheds Technical Release 55 (TR-55)", June 1986, \url{https://web.archive.org/web/20230810204711/https://directives.sc.egov.usda.gov/OpenNonWebContent.aspx?content=22162.wba} [Recovered with the Internet Archive: Wayback Machine]
}
\usage{
Ia
}
\description{
A table containing curve numbers for initial abstraction.
}
\keyword{datasets}
