% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iemiscdata.R
\docType{package}
\name{iemiscdata-package}
\alias{iemiscdata}
\alias{iemiscdata-package}
\title{iemiscdata: Irucka Embry's Miscellaneous Data Collection}
\description{
iemiscdata provides an eclectic set of data sets [Chemistry, Engineering
Economics, Environmental/Water Resources Engineering, Nuclear Accidents, US
Presidential Elections, and US Continental Congress Presidents].
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/iembry/iemiscdata}
  \item Report bugs at \url{https://gitlab.com/iembry/iemiscdata/-/issues}
}

}
\author{
\strong{Maintainer}: Irucka Embry \email{iembry@ecoccs.com}

}
\keyword{internal}
