% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iemiscdata.R
\docType{package}
\name{iemiscdata}
\alias{iemiscdata}
\alias{iemiscdata-package}
\title{iemiscdata: Irucka Embry's miscellaneous data collection}
\description{
iemiscdata provides an eclectic set of data sets [Engineering Economics,
  Environmental/Water Resources Engineering, US Presidential Elections,
  etc.].
}

