% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-military_nuclear_accidents_wiki.R
\docType{data}
\name{military_nuclear_accidents_wiki}
\alias{military_nuclear_accidents_wiki}
\title{Military nuclear accidents (Wikipedia)}
\format{
A data.table data frame with 76 rows and 4 variables:
\describe{
\item{Date}{Date of the accident}
\item{Location}{Location of the accident}
\item{Type}{Type of accident}
\item{Description}{Description of accident or incident}
}
}
\source{
Wikimedia Foundation, Inc. Wikipedia, 15 September 2023, "List of military nuclear accidents", \url{https://en.wikipedia.org/wiki/List_of_military_nuclear_accidents}.
}
\usage{
military_nuclear_accidents_wiki
}
\description{
A table containing the military nuclear accidents.
}
\keyword{datasets}
