% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-nsheetflow.R
\docType{data}
\name{nsheetflow}
\alias{nsheetflow}
\title{Table 3-1: Roughness coefficients (Manning's n) for sheet flow}
\format{
A data.table data frame with 12 rows and 2 variables:
\describe{
\item{Surface description}{Description of the surface}
\item{"n *1}{Manning's roughness coefficient}
}
}
\source{
United States Department of Agriculture Natural Resources Conservation Service Conservation Engineering Division, "Urban Hydrology for Small Watersheds Technical Release 55 (TR-55)", June 1986, page 3-3, \url{https://web.archive.org/web/20230810204711/https://directives.sc.egov.usda.gov/OpenNonWebContent.aspx?content=22162.wba} [Recovered with the Internet Archive: Wayback Machine]
}
\usage{
nsheetflow
}
\description{
A table containing Manning's roughness coefficient for sheet flow.
}
\keyword{datasets}
