% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-nsheetflow_notes.R
\docType{data}
\name{nsheetflow_notes}
\alias{nsheetflow_notes}
\title{Notes for Table 16-2 Computation of coordinates for unit hydrograph for use in example 16-1}
\format{
A data.table data frame with 3 rows and 2 variables:
\describe{
item{Note Number (*)}{The note numbers}
item{Notes}{The notes}
}
}
\source{
United States Department of Agriculture Natural Resources Conservation Service Conservation Engineering Division, "Urban Hydrology for Small Watersheds Technical Release 55 (TR-55)", June 1986, page 3-3, \url{https://directives.sc.egov.usda.gov/OpenNonWebContent.aspx?content=22162.wba}
}
\usage{
nsheetflow_notes
}
\description{
A table containing the notes for Manning's roughness coefficient for sheet
flow.
}
\keyword{datasets}
