% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-nuclear_accidents.R
\docType{data}
\name{nuclear_accidents}
\alias{nuclear_accidents}
\title{Nuclear & Radiation Accidents and incidents (The Guardian)}
\format{
A data.table data frame with 33 rows and 6 variables:
\describe{
\item{Date}{Year of the accident}
\item{Incident}{Name of the incident}
\item{INES rating}{INES level}
\item{Country}{Country where the accident took place}
\item{Location}{Location of the accident}
\item{IAEA description}{IAEA description of accident or incident}
}
}
\source{
The Guardian, "Nuclear power plant accidents: listed and ranked since 1952: How many nuclear power plants have had accidents and incidents?", \url{https://www.theguardian.com/news/datablog/2011/mar/14/nuclear-power-plant-accidents-list-rank}.
}
\usage{
nuclear_accidents
}
\description{
A table containing nuclear & radiation accidents and incidents since 1952.
}
\keyword{datasets}
