% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-nuclear_accidents_ranked.R
\docType{data}
\name{nuclear_accidents_ranked}
\alias{nuclear_accidents_ranked}
\title{Ranked Nuclear & Radiation Accidents and incidents (The Guardian)}
\format{
A data.table data frame with 7 rows and 6 variables:
\describe{
\item{Level}{Ranked level of severity of the accident}
\item{Definition}{Definition associated with each level}
\item{People and environment}{Any harm to Human Beings and/or the surrounding environment}
\item{Radiological barriers & control}{Any radiological barriers and/or controls}
\item{Defence in depth}{Defensive barriers described}
\item{Example}{An example of an accident or incident at the given level}
}
}
\source{
The Guardian, "Nuclear power plant accidents: listed and ranked since 1952: How many nuclear power plants have had accidents and incidents?", \url{https://www.theguardian.com/news/datablog/2011/mar/14/nuclear-power-plant-accidents-list-rank}.
}
\usage{
nuclear_accidents_ranked
}
\description{
A table containing the military nuclear accidents.
}
\keyword{datasets}
