% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-nuclear_accidents_wiki.R
\docType{data}
\name{nuclear_accidents_wiki}
\alias{nuclear_accidents_wiki}
\title{Nuclear and radiation accidents and incidents (Wikipedia)}
\format{
A data.table data frame with 28 rows and 6 variables:
\describe{
\item{Date}{Date of the accident}
\item{Location}{Location of the accident}
\item{Description}{Description of accident or incident}
\item{Fatalities}{Number of people that died}
\item{Cost (in millions 2006 US$)}{Cost in millions of 2006 US dollars}
\item{INES rating)}{INES level}
}
}
\source{
Wikimedia Foundation, Inc. Wikipedia, 5 September 2023, "Nuclear and radiation accidents and incidents", \url{https://en.wikipedia.org/wiki/Nuclear_and_radiation_accidents_and_incidents}.
}
\usage{
nuclear_accidents_wiki
}
\description{
A table containing nuclear and radiation accidents and incidents.
}
\keyword{datasets}
