% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-nuclear_power_accidents_country_wiki.R
\docType{data}
\name{nuclear_power_accidents_country_wiki}
\alias{nuclear_power_accidents_country_wiki}
\title{Nuclear power accidents by country (Wikipedia)}
\format{
A data.table data frame with 133 rows and 9 variables:
\describe{
\item{Date}{Date of the accident}
\item{Country}{Country where the accident took place}
\item{Location}{Location of the accident}
\item{Description}{Description of accident or incident}
\item{INES rating)}{INES level}
\item{Fatalities}{Number of people that died}
\item{Victims}{Number of people that had adverse health effects, but did not die}
\item{Cost (in millions 2006 US$)}{Cost in millions of 2006 US dollars}
\item{Cost}{Cost}
}
}
\source{
Wikimedia Foundation, Inc. Wikipedia, 24 June 2024, "List of nuclear power accidents by country", \url{https://en.wikipedia.org/wiki/List_of_nuclear_power_accidents_by_country}.
}
\usage{
nuclear_power_accidents_country_wiki
}
\description{
A table containing the nuclear power accidents by country.
}
\keyword{datasets}
