% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-plant_essentials.R
\docType{data}
\name{plant_essentials}
\alias{plant_essentials}
\title{Table of Plant Nutrients}
\format{
A data.table data frame with 11 rows and 2 variables:
\describe{
\item{Elemental Macronutrients}{The plant macronutrients}
\item{Elemental Micronutrients}{The plant micronutrients}
}
}
\source{
Irucka Ajani Embry, "Table of Plant Nutrients", \emph{Tuning to the Soil-ed Seed Vibrations For Our Collective Health}, Nashville, Tennessee: Questioning the Universe Publishing, 2024, page 15. See \url{https://www.qup.quest/soiled-seeds.html}.
}
\usage{
plant_essentials
}
\description{
A table containing the plant macro- and micro-nutrients as chemical elements.
This data set contains the content of "Table of Plant Nutrients" without the
element oxidation states.
}
\keyword{datasets}
