% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pres_elect1.R
\docType{data}
\name{pres_elect1}
\alias{pres_elect1}
\title{Wikipedia Table of the United States of America Presidential Elections}
\format{A data.table data frame with 57 rows and 25 variables:
\describe{
\item{Election year}{Year(s) of the election}
\item{Election year begin}{The year the election began}
\item{Election year end}{The year the election ended}
\item{Winner}{Winner's name}
\item{Winner Party}{Winner's political party}
\item{Winner Electoral Votes 1}{Number of winner's electoral votes}
\item{Winner Electoral Votes 2}{Number of winner's electoral votes}
\item{Major Candidate1}{Candidate1's name}
\item{Major Candidate1 Party}{Candidate1's political party}
\item{Major Candidate1 Electoral Votes}{Number of Candidate1's electoral
     votes}
\item{Major Candidate2}{Candidate2's name}
\item{Major Candidate2 Party}{Candidate2's political party}
\item{Major Candidate2 Electoral Votes}{Number of Candidate2's electoral
     votes}
\item{Major Candidate3}{Candidate3's name}
\item{Major Candidate3 Party}{Candidate3's political party}
\item{Major Candidate3 Electoral Votes}{Number of Candidate3's electoral
     votes}
\item{Major Candidate4}{Candidate4's name}
\item{Major Candidate4 Party}{Candidate4's political party}
\item{Major Candidate4 Electoral Votes}{Number of Candidate4's electoral
     votes}
\item{Major Candidate5}{Candidate5's name}
\item{Major Candidate5 Party}{Candidate5's political party}
\item{Major Candidate5 Electoral Votes}{Number of Candidate5's electoral
     votes}
\item{Major Candidate6}{Candidate6's name}
\item{Major Candidate6 Party}{Candidate6's political party}
\item{Major Candidate6 Electoral Votes}{Number of Candidate6's electoral
     votes}
}}
\usage{
pres_elect1
}
\description{
A table containing the results of the US Presidential Elections.
}
\details{
"Notes regarding the special characters used in the table:
* Winner received less than an absolute majority of the popular vote.
}
\examples{
pres_elect1
}
\references{
Wikimedia Foundation, Inc. Wikipedia, 10 February 2016, “United States presidential election”, \url{https://en.wikipedia.org/wiki/United_States_presidential_election}.
}

