% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-r20.R
\docType{data}
\name{r20}
\alias{r20}
\title{20 Percent Nominal Interest Table (Engineering Economy)}
\format{
A data.table data frame with 44 rows and 6 variables:
\describe{
\item{n}{The "number of periods (years)"}
\item{F/P}{Future value given present value}
\item{P/F}{Present value given future value}
\item{F/A}{Future value given annual value}
\item{P/A}{Present value given annual value}
\item{n}{The "number of periods (years)"}
}
}
\source{
William G. Sullivan, Elin M. Wicks, and C. Patrick Koelling, \emph{Engineering Economy}, Fourteenth Edition, Upper Saddle River, New Jersey: Pearson/Prentice Hall, 2009, page 625-647.
}
\usage{
r20
}
\description{
A table containing the number of periods and the corresponding engineering
economic value given the "nominal annual interest rate, compounded
continuously" and period.
}
\keyword{datasets}
