% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-r8.R
\docType{data}
\name{r8}
\alias{r8}
\title{8 Percent Nominal Interest Table (Engineering Economy)}
\format{A data.table data frame with 44 rows and 6 variables:
\describe{
\item{n}{The "number of periods (years)"}
\item{F/P}{Future value given present value}
\item{P/F}{Present value given future value}
\item{F/A}{Future value given annual value}
\item{P/A}{Present value given annual value}
\item{n}{The "number of periods (years)"}
}}
\usage{
r8
}
\description{
A table containing the number of periods and the corresponding engineering
economic value given the "nominal annual interest rate, compounded
continuously" and period.
}
\examples{
r8
}
\references{
William G. Sullivan, Elin M. Wicks, and C. Patrick Koelling, \emph{Engineering Economy}, Fourteenth Edition, Upper Saddle River, New Jersey: Pearson/Prentice Hall, 2009, page 625-647.
}

