% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/data-rainfall_distribution_equation_coefficients.R
\docType{data}
\name{rainfall_distribution_equation_coefficients}
\alias{rainfall_distribution_equation_coefficients}
\title{Table F-2 Coefficients for the equation used to generate figure 6-1}
\format{
A data.table data frame with 4 rows and 5 variables:
\describe{
\item{Rainfall distribution}{Rainfall distribution type}
\item{C0}{Coefficient value}
\item{C1}{Coefficient value}
\item{C2}{Coefficient value}
\item{C3}{Coefficient value}
}
}
\source{
United States Department of Agriculture Natural Resources Conservation Service Conservation Engineering Division, "Urban Hydrology for Small Watersheds Technical Release 55 (TR-55)", June 1986, \url{https://web.archive.org/web/20230810204711/https://directives.sc.egov.usda.gov/OpenNonWebContent.aspx?content=22162.wba} [Recovered with the Internet Archive: Wayback Machine]
}
\usage{
rainfall_distribution_equation_coefficients
}
\description{
A table containing the coefficients to generate Figure 6-1.
}
\keyword{datasets}
