% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-rainfall_type_equation_coefficients.R
\docType{data}
\name{rainfall_type_equation_coefficients}
\alias{rainfall_type_equation_coefficients}
\title{Table F-1 Coefficients for the equation used to generate exhibits 4-I through 4-III}
\format{
A data.table data frame with 28 rows and 5 variables:
\describe{
\item{Rainfall type}{Rainfall distribution type}
\item{Ia/P}{"Initial abstraction" over Rainfall ratio}
\item{C0}{Coefficient value}
\item{C1}{Coefficient value}
\item{C2}{Coefficient value}
}
}
\source{
United States Department of Agriculture Natural Resources Conservation Service Conservation Engineering Division, "Urban Hydrology for Small Watersheds Technical Release 55 (TR-55)", June 1986, \url{https://directives.sc.egov.usda.gov/OpenNonWebContent.aspx?content=22162.wba}
}
\usage{
rainfall_type_equation_coefficients
}
\description{
A table containing the coefficients to generate exhibits 4-1 to 4-III.
}
\keyword{datasets}
