% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-riprap_table_ft.R
\docType{data}
\name{riprap_table_ft}
\alias{riprap_table_ft}
\title{Tennessee Department of Transportation (TDOT) Machined riprap Table (feet)}
\format{
A data.table data frame with 5 rows and 4 variables:
\describe{
\item{Machined riprap Class}{TDOT class of machined riprap}
\item{Minimum Diameter Size [ft]}{Minimum diameter size of riprap}
\item{Average Diameter Size [ft]}{Average diameter size of riprap}
\item{Maximum Diameter Size [ft]}{Maximum diameter size of riprap}
}
}
\source{
\enumerate{
   \item Tennessee Department of Transportation (TDOT) 2015 Standard Specifications, \url{https://www.tn.gov/content/tn/tdot/tdot-construction-division/transportation-construction-division-resources/transportation-construction-2015-standard-specifications.html}.
   \item Tennessee Department of Transportation (TDOT) Standard Specifications for Road and Bridge Construction, \url{https://www.tn.gov/content/dam/tn/tdot/construction/old_web_page/TDOT_2015_Spec_Book_FINAL_pdf.pdf}.
   \item City of Knoxville BMP Manual Erosion & Sediment -- ACTIVITY: Riprap ES–23, May 2003 \url{https://cdnsm5-hosted.civiclive.com/UserFiles/Servers/Server_109478/File/Engineering/BMPManual/ES-23.pdf}.
}
}
\usage{
riprap_table_ft
}
\description{
A table containing the sizes of TDOT machined riprap in feet.
}
\keyword{datasets}
