% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-uh_ratios.R
\docType{data}
\name{uh_ratios}
\alias{uh_ratios}
\title{Table 16-2 Computation of coordinates for unit hydrograph for use in example 16-1}
\format{
A data.table data frame with 51 rows and 2 variables:
\describe{
\item{Time ratios (t/Tp)}{"ratio of the time of DUH coordinate to time to peak of the DUH"}
\item{Discharge ratios (q/qp)}{"ratio of discharge at a certain time to the peak discharge of the unit hydrograph (UH)"}
}
}
\source{
United States Department of Agriculture (USDA) Natural Resources Conservation Service (NRCS), \emph{National Engineering Handbook}, "Part 630 Hydrology Chapter 16 Hydrographs", Pages 16-4, 16-9, 16-15, \url{https://web.archive.org/web/20230615195809/https://directives.sc.egov.usda.gov/OpenNonWebContent.aspx?content=17755.wba} [Recovered with the Internet Archive: Wayback Machine]
}
\usage{
uh_ratios
}
\description{
A table containing the coordinates for a unit hydrograph. Developed for use
in Example 16-1.
}
\keyword{datasets}
