% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-USA_primary_water_contaminants_notes.R
\docType{data}
\name{USA_primary_water_contaminants_notes}
\alias{USA_primary_water_contaminants_notes}
\title{US EPA National Primary Drinking Water Regulations Contaminants Notes}
\format{
A data.table data frame with 29 rows and 2 variables:
\describe{
\item{Note Number (*)}{The note numbers}
\item{Notes}{The notes}
}
}
\source{
United States (US) Environmental Protection Agency (EPA): "National Primary Drinking Water Regulations", \url{https://web.archive.org/web/20230801024443/https://www.epa.gov/ground-water-and-drinking-water/national-primary-drinking-water-regulations}. Used the Internet Archive: Wayback Machine archived version for acceptance into CRAN.
}
\usage{
USA_primary_water_contaminants_notes
}
\description{
A table of notes for the primary drinking water pollutants table.
}
\keyword{datasets}
