% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-USA_secondary_water_contaminants.R
\docType{data}
\name{USA_secondary_water_contaminants}
\alias{USA_secondary_water_contaminants}
\title{US EPA Secondary Drinking Water Standards}
\format{
A data.table data frame with 15 rows and 3 variables:
\describe{
\item{Contaminant}{The name of the contaminant}
\item{Secondary MCL}{Maximum Contaminant Level (MCL)}
\item{Noticeable Effects above the Secondary MCL}{What are the noticeable effects above the secondary MCL}
}
}
\source{
United States (US) Environmental Protection Agency (EPA): "Secondary Drinking Water Standards: Guidance for Nuisance Chemicals", \url{https://web.archive.org/web/20230903054648/https://www.epa.gov/sdwa/secondary-drinking-water-standards-guidance-nuisance-chemicals}. Used the Internet Archive: Wayback Machine archived version for acceptance into CRAN.
}
\usage{
USA_secondary_water_contaminants
}
\description{
A table containing the secondary drinking water pollutants.
}
\keyword{datasets}
