% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-USA_secondary_water_contaminants_notes.R
\docType{data}
\name{USA_secondary_water_contaminants_notes}
\alias{USA_secondary_water_contaminants_notes}
\title{US EPA Secondary Drinking Water Standards Notes}
\format{
A data.table data frame with 1 rows and 2 variables:
\describe{
\item{Note Number (*)}{The note numbers}
\item{Notes}{The notes}
}
}
\source{
United States (US) Environmental Protection Agency (EPA): "Secondary Drinking Water Standards: Guidance for Nuisance Chemicals", \url{https://web.archive.org/web/20230903054648/https://www.epa.gov/sdwa/secondary-drinking-water-standards-guidance-nuisance-chemicals}. Used the Internet Archive: Wayback Machine archived version for acceptance into CRAN.
}
\usage{
USA_secondary_water_contaminants_notes
}
\description{
A table of notes for the secondary drinking water pollutants table.
}
\keyword{datasets}
