% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-water_vapour_pressure.R
\docType{data}
\name{water_vapour_pressure}
\alias{water_vapour_pressure}
\title{Vapour Pressure of Water (Wikipedia)}
\format{
A data.table data frame with 21 rows and 5 variables:
\describe{
\item{T, degrees C}{Temperature in degrees Celsius}
\item{T, degrees F}{Temperature in degrees Fahrenheit}
\item{P, kPa}{Pressure in kilo Pascals}
\item{P, torr}{Pressure in Torricelli}
\item{P, atm}{Pressure in atmospheres}
}
}
\source{
Wikimedia Foundation, Inc. Wikipedia, 6 September 2023, "Vapour pressure of water", \url{https://en.wikipedia.org/wiki/Vapour_pressure_of_water}.
}
\usage{
water_vapour_pressure
}
\description{
A table containing the vapor pressure of water at different temperatures &
pressures.
}
\keyword{datasets}
