% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-greenhouse_gases_contributions.R
\docType{data}
\name{greenhouse_gases_contributions}
\alias{greenhouse_gases_contributions}
\title{Greenhouse Gases -- Percent Contribution to Total Greenhouse Effect (NASA)}
\format{
A data.table data frame with 4 rows and 3 variables:
\describe{
\item{Gas_Name}{Name of the gas}
\item{Gas_Formula}{Chemical formula of the gas}
\item{Greenhouse Gas Contributions (approximate percent values)}{Individual greenhouse contribution by percent}
}
}
\source{
NASA Goddard Space Flight Center: EOS Project Science Office: Earth Observatory. Jun 16, 2011, "Effects of Changing the Carbon Cycle", \url{https://earthobservatory.nasa.gov/features/CarbonCycle/page5.php}.
}
\usage{
greenhouse_gases_contributions
}
\description{
A table containing the percent contribution to the total greenhouse effect.
}
\keyword{datasets}
